Removed progress HUD. It turned out the only reason it was need was a bad algorithm on my part. Modified a couple of DDMathParser’s files to silence warnings. Explicit cast to long was required to shut it up…

+++ Reimann Sums/DDMathStringToken.m
@@ -78,7 +78,8 @@
         [d appendFormat:@"arity:%@, ", arityNames[arity]];
         
         NSInteger precedence = [self operatorPrecedence];
-        [d appendFormat:@"precedence:%ld, ", precedence];
+        [d appendFormat:@"precedence:%ld, ", (long)
+		 precedence];
         
         DDOperatorAssociativity assoc = [self operatorAssociativity];
         NSString *assocNames[2] = { @"LEFT", @"RIGHT" };
